Move flexible types to dedicated file
